Add meta tag to implement Bootstrap 3's mobile-first design
Bootstrap 3 is designed to be responsive to mobile devices. Mobile-first
styles are part of the core framework. Hence, the following meta tag and
its attributes were added:

1. width=device-width
Sets the width of the page to follow the screen-width of the device
(which will vary depending on the device).

2. initial-scale=1 part
Sets the initial zoom level when the page is first loaded by the
browser.